+++ source/components/input/story.js
@@ -30,23 +30,40 @@ storiesOf('Input', module)
       </Row>
     )),
   )
-  .add('with single error', () =>
+  .add('with description', () =>
     ['input', 'textarea'].map((element) => (
       <Row key={element}>
         <Input.Uncontrolled
           element={element}
           label="E-mail"
-          errors="Incorrect e-mail"
+          description="example@example.com"
         />
       </Row>
     )),
   )
-  .add('with multiple errors', () =>
+  .add('with errors', () =>
     ['input', 'textarea'].map((element) => (
       <Row key={element}>
         <Input.Uncontrolled
           element={element}
           label="E-mail"
+          errors={[
+            'must not be empty',
+            'must contain @ character',
+            'must be at least 3 characters long',
+          ]}
+        />
+      </Row>
+    )),
+  )
+  .add('with placeholder, description and errors', () =>
+    ['input', 'textarea'].map((element) => (
+      <Row key={element}>
+        <Input.Uncontrolled
+          element={element}
+          label="E-mail"
+          placeholder="example@example.com"
+          description="Your e-mail"
           errors={[
             'must not be empty',
             'must contain @ character',